+++ src/auth.ts
@@ -24,19 +24,27 @@ export async function ensureAuthenticated() {
 
     const response = await window.showErrorMessage("You need to sign-in with GitHub to perform this operation.", TOKEN_RESPONSE);
     if (response === TOKEN_RESPONSE) {
-        const value = await env.clipboard.readText();
-        const token = await window.showInputBox({ prompt: "Enter your GitHub token", value });
-        if (token) {
-            await keytar.setPassword(SERVICE, ACCOUNT, token);
-        } else {
-            throw new Error("Authentication required")
-        }
+        await signIn();
     }
+}
+
+export async function isAuthenticated() {
+    return await getToken() !== null;
 }
 
 export async function getToken() {
 	const token = await keytar.getPassword(SERVICE, ACCOUNT);
 	return token;
+}
+
+export async function signIn() {
+    const value = await env.clipboard.readText();
+    const token = await window.showInputBox({ prompt: "Enter your GitHub token", value });
+    if (token) {
+        await keytar.setPassword(SERVICE, ACCOUNT, token);
+    } else {
+        throw new Error("Authentication required");
+    }
 }
 
 export async function signout() {