+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## v0.0.4 (11/27/2016)
 
-* Allow seeding multiple files in a new gist, by specifying a comma-seperated list of files names
-* Update the behavior of opening gists, to open them as "loose files" instead of as a workspace
+* Improved the `GistFS: Delete Gist` command, by allowing you to select a Gist from your list of Gists, instead of just providing a Gist ID
+* Improved the `GistFS: Open Gist` command, by allowing you to select one of your Gists from a list, instead of just providing a Gist ID
+* Updated the default behavior of opening gists, to open them as "loose files" instead of as a workspace
+* Introduced a new `GistFS: Open Gist as Workspace` command, which behaves the same as `GistFS: Open Gist`, but opens the Gist as a workspace instead of "loose files"
+* Added support for being able to seed multiple files when creating a Gist, by specifying a comma-seperated list of files names
+* Added a progress indicator when creating a new Gist
+* Introduced a new `GistFS: Sign In` command, to explicitly sign in to your GitHub account
 
 ## v0.0.3 (11/26/2016)